+++ libs/react-components/src/lib/components/button/changelog.md
@@ -2,5 +2,6 @@
 
 ## 0.1.0
 
+- [MINOR] _Improved_: Added support for utilising `button` styles with HTML anchor elements
 - [MINOR] _Improved_: Added ref prop to enable consuming applications to reference HTML button element
 - [MINOR] Initial component creation. Support for primary and secondary variants and disabled state.